fix(typo): on :feature:profile-details' package name
