+++ app/src/main/java/com/jeanbarrossilva/mastodonte/app/feature/profiledetails/ProfileModule.kt
@@ -2,7 +2,7 @@ package com.jeanbarrossilva.mastodonte.app.feature.profiledetails
 
 import androidx.annotation.IdRes
 import androidx.fragment.app.FragmentManager
-import com.jeanbarrossilva.mastodon.feature.profiledetails.ProfileDetailsBoundary
+import com.jeanbarrossilva.mastodonte.feature.profiledetails.ProfileDetailsBoundary
 import org.koin.core.module.Module
 import org.koin.dsl.module
