+++ feature/profile-details/src/androidTest/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <application>
         <activity
-            android:name="com.jeanbarrossilva.mastodon.feature.profiledetails.test.ProfileDetailsActivity" />
+            android:name="com.jeanbarrossilva.mastodonte.feature.profiledetails.test.ProfileDetailsActivity" />
     </application>
 </manifest>
