--- feature/profile-details/src/androidTest/java/com/jeanbarrossilva/mastodon/feature/profiledetails/test/TestProfileDetailsBoundary.kt
@@ -1,12 +1,0 @@
-package com.jeanbarrossilva.mastodon.feature.profiledetails.test
-
-import com.jeanbarrossilva.mastodon.feature.profiledetails.ProfileDetailsBoundary
-import java.net.URL
-
-internal class TestProfileDetailsBoundary : ProfileDetailsBoundary {
-    override fun navigateToWebpage(url: URL) {
-    }
-
-    override fun navigateToTootDetails(id: String) {
-    }
-}
