+++ feature/profile-details/src/androidTest/java/com/jeanbarrossilva/mastodonte/feature/profiledetails/test/TestProfileDetailsBoundary.kt
@@ -1,0 +1,12 @@
+package com.jeanbarrossilva.mastodonte.feature.profiledetails.test
+
+import com.jeanbarrossilva.mastodonte.feature.profiledetails.ProfileDetailsBoundary
+import java.net.URL
+
+internal class TestProfileDetailsBoundary : ProfileDetailsBoundary {
+    override fun navigateToWebpage(url: URL) {
+    }
+
+    override fun navigateToTootDetails(id: String) {
+    }
+}
