--- feature/profile-details/src/main/java/com/jeanbarrossilva/mastodon/feature/profiledetails/viewmodel/ContextProvider.kt
@@ -1,9 +1,0 @@
-package com.jeanbarrossilva.mastodon.feature.profiledetails.viewmodel
-
-import android.content.Context
-
-/** Provides a [Context] through [provide]. **/
-internal fun interface ContextProvider {
-    /** Provides a [Context]. **/
-    fun provide(): Context
-}
